Block: combine store subscriptions
