+++ packages/block-editor/src/components/block-edit/index.js
@@ -1,19 +1,8 @@
 /**
- * WordPress dependencies
- */
-import { useMemo } from '@wordpress/element';
-
-import { hasBlockSupport } from '@wordpress/blocks';
-/**
  * Internal dependencies
  */
 import Edit from './edit';
-import {
-	BlockEditContextProvider,
-	useBlockEditContext,
-	mayDisplayControlsKey,
-	mayDisplayParentControlsKey,
-} from './context';
+import { useBlockEditContext } from './context';
 
 /**
  * The `useBlockEditContext` hook provides information about the block this hook is being used in.
@@ -25,54 +14,6 @@ import {
  */
 export { useBlockEditContext };
 
-export default function BlockEdit( {
-	mayDisplayControls,
-	mayDisplayParentControls,
-	// The remaining props are passed through the BlockEdit filters and are thus
-	// public API!
-	...props
-} ) {
-	const {
-		name,
-		isSelected,
-		clientId,
-		attributes = {},
-		__unstableLayoutClassNames,
-	} = props;
-	const { layout = null } = attributes;
-	const layoutSupport =
-		hasBlockSupport( name, 'layout', false ) ||
-		hasBlockSupport( name, '__experimentalLayout', false );
-	return (
-		<BlockEditContextProvider
-			// It is important to return the same object if props haven't
-			// changed to avoid  unnecessary rerenders.
-			// See https://reactjs.org/docs/context.html#caveats.
-			value={ useMemo(
-				() => ( {
-					name,
-					isSelected,
-					clientId,
-					layout: layoutSupport ? layout : null,
-					__unstableLayoutClassNames,
-					// We use symbols in favour of an __unstable prefix to avoid
-					// usage outside of the package (this context is exposed).
-					[ mayDisplayControlsKey ]: mayDisplayControls,
-					[ mayDisplayParentControlsKey ]: mayDisplayParentControls,
-				} ),
-				[
-					name,
-					isSelected,
-					clientId,
-					layoutSupport,
-					layout,
-					__unstableLayoutClassNames,
-					mayDisplayControls,
-					mayDisplayParentControls,
-				]
-			) }
-		>
-			<Edit { ...props } />
-		</BlockEditContextProvider>
-	);
+export default function BlockEdit( props ) {
+	return <Edit { ...props } />;
 }
